Changed window_manager to db_app, moved run function from add_data to db_app. Now login happens withing gui as well!
